Multiple hostnames
Change the cli to accept multiple hostnames on the command line. Also,
in signer, the Hosts list was moved out of the Subject field of Certificate
Sign Request, and put into the CSR struct itself, and the Hostname field was
removed. Any part of the code that manually built a CSR struct had to
therefore be changed.

NOTE: The bundler test TCP dial error is still occuring when I run from my home
wifi. I'm committing anyway on the basis of that being out of my hands.