Add support for scanning server elliptic curve preferences.
This change adds support for scanning server curve choice preferences for the ECDH cipher suites, which is interesting in light of logjam (even though that only affects non-ED DH, it's nice to know what's happening in case something similar becomes interesting in the future). This fixes #221, although it's slightly more involved than @jacobhaven suggested. Specifically, the server curve preference is not declared until the ServerKeyExchange message, meaning that you have to go well past ServerHello to extract it.

The scan is performed using the same algorithm as for cipher suites, where the server's negotiated preference is dropped iteratively from the list of all possible choices until we've enumerated the server's preference list.