Improvement: Adds new cli flag 'int-dir' to bundle cmd
The bundle command creates a special intermediate folder to store
the downloaded intermediate certificates.
The current implementation creates an 'intermediates' folder
at the current path, where cfssl is executed. This polutes the
file system with many unnecessary 'intermediates' folders.
This commit fixes the problem by adding a new flag 'int-dir' to
customize the intermediate folder path. The same flag is already
used for the HTTP server and inherits it's default value from
the config.