Set pkcs11 user pin in environment
To hide the user pin from the process list, set it in the environment. That way, only users with access to `cat /proc/$(pidof cfssl)/environ` (typically root) and  can read the user pin. You can store the pin in a file accessible by root only and set the pin as follows:
````USER_PIN=$(cat /etc/cfssl/pin)
     $GOPATH/bin/cfssl ...

If USER_PIN is unset, the behavior is identical to the current one (i.e. you need to provide a pin on the command line).