Apply label to allow containerized build under SELinux
Under the default SELinux policy, Docker containers can only access
files labeled svirt_sandbox_file_t and optionally the right MCS tags.
The [Zz] mount options relabel to fix this, but are new in 1.7 and break
RW mounting in prior versions. This is an alternative attempt at
allowing the build container to work.