Added CSR functionality.
The `csr` package currently only operates on self-contained
CertificateRequest strucutres. There are times, however, where it is
useful to generate a CSR with an existing key or to re-sign a CSR with
an existing private key. This is mostly useful for client tools that
want to make use of the CFSSL packages.

This pull request provides two new functions:

+ Regenerate, which re-signs a CSR with a private key, and

+ Generate, which uses a CertificateRequest to build a CSR that is
  signed with a provided private key.

To support this, some additional functions were added:

+ csr.New provides a convenience method for creating a new
  CertificateRequest. The changes to KeyRequest means that both

```
  var req csr.CertificateRequest
  err = json.Unmarshal(in, &req)
```

  and

```
  var req = &csr.CertificateRequest{}
  err = json.Unmarshal(in, &req)
```

  don't work anymore. New provides a way to do this by setting
  up a new CertificateRequest with a BasicKeyRequest.

+ helpers.ParseCSR can parse a CSR. It checks the signature on the
  CSR.

+ helpers.SignerAlgo takes a crypto.Signer and crypto.Hash and
  returns the appropriate signature algorithm.