Reject keys with CKA_ALWAYS_AUTHENTICATE.
If the private key has a CKA_ALWAYS_AUTHENTICATE attribute, it means we have to
Login before each sign attempt, and that would be terribly slow.

Note: This means that a default-setup Yubikey Neo in PIV mode will be rejected.
I'm planning to see if it's possible to set the attribute to false.