Ignore unsupported CSR extensions.
Previously, we would hard fail on any CSR extensions that weren't whitelisted.
However, as Hugo pointed out in
https://github.com/letsencrypt/boulder/pull/1224#issuecomment-174065959,
this will break a large number of clients, which include extensions in the CSR
that we don't explicitly process. The new behavior is to copy only the
whitelisted extensions into the certificate, and ignore the rest.