Fix broken cache-control logic
The previous patch refactoring cache-control semantics attempted
to add the headers to the response after a WriteHeader or Write
call which prevents additional headers being written to the response
unless they are specified as trailers (which they weren't). This
simplifies the logic and moves the header add/set calls before
any response WriteHeader/Write calls.