Add basic ETag support
RFC 5019 (Lightweight OCSP profile) recommends that responders
implement the ETag header in responses and to properly behave
when If-None-Match header is sent in requests. This implementation
simply sets the ETag to the SHA256 hash of the OCSP response returned
from a Source. If a If-None-Match header is sent the Source will be
queried and if the hashed response matches the header a 304 Not Modified
header will be returned to the client instead of a 200 and the response
body.