Remove data race in initca.
The default policy was previously a global variable, which caused
issues when multiple CAs were being initialised in parallel (such
as in testing, e.g. [1]).

This change causes the default policy to be a function returning a
default policy. The function can still be swapped out in testing as
needed, but the functions now operate on local copies of the default
policy. A change to the default policy (which is only used in testing)
can now be accomplished by changing the function itself.

Furthermore, based on comments from Jacob Hoffman-Andrews (@jsha), the
signer package now uses a local issuerURL variable.

In an effort to help catch these issues in the future, the race
detector is now being run by Travis.

[1] https://github.com/cloudflare/cfssl/issues/571