Add ability to override validity period on a per-cert basis.
In some cases one needs to set the validity period (notBefore, notAfter) to
an exact time, e.g. when a certificate's validity period needs to exactly
match another certificate's validity period. In particular, if one is
generating CT precertificates and the final certificate using separate
invocations of the Signer, then the certificate's validity period must be
set to exactly the same values as the precertificate's validity period.