Performance optimization for local Signer Sign()
When the local signer does not have a dbAccessor set, It does not need to get a AuthorityKeyId (the only reason ParseCertificatePEM is called). ParseCertificatePEM is expensive with ECC since it involves doing a verification. Moving this call into the if statement means that the Sign function runs nearly twice as fast on my machine when no dbAccessor is set (it likely doesn't matter as much for RSA since verification on RSA is relatively cheap).