Fix remaining bundler tests
Well, "fix." The failing test was ultimately disabled b/c it was
crafted for the exact certificate chain returned by a site which no
longer exists, and I haven't been able to find a suitable replacement.
However, other previously disabled tests were re-enabled, and the
overall coverage for the functionality is pretty good.