Fix refusing to parse cert with multiple PEM block
The original behaviour of cfssl is to reject parsing the pem file if
there are more than 1 valid PEM blocks. This commit adds the capability
of selecting the first valid PEM certificate block from the file and
then display results.

The new code will keep looking through the valid PEM blocks and find the
one with a `CERTIFICATE` type and use it. If the `rest` value is not
`nil`, i.e. there are blocks left, it will show up an info about that
and continue.

Since, the changes are only being used by the `cfssl certinfo -cert`
command, tests have been added to verify this commands behaviour.