Handle reading from Windows filepaths
In helpers.GetBytes, absolute paths starting with a Windows
drive letter cause the function to attempt to parse the drive
letter as a prefix.  For instance, on input such as `c:\certs\file`
the function would return an error `unknown prefix: c`.

Change this so that the function will attempt to use the OS independent
methods to read the file if the prefix is not explicitly file or env.