Add a PKCS#11 support for multirootca command.
Add recognition of a `pkcs11` protocol in the URI parsing for the private key in the multirootca
config. Add other configuration options when this is used in order to identify the correct token
for the server to use, and then return a crypto.Signer that uses a Private Key located on the
token by its ID. Also add recognition for crypto11.PKCS11PrivateKeyRSA and
crypto11.PKCS11PrivateKeyECDSA in the multirootca binary as valid private key types.

Signed-off-by: Danny Wensley <danny@wensley.eu>