Convert KeyRequest into a concrete type to fix yaml support.
The KeyRequest interface isn't providing any direct value to the
current code, but it explicitly breaks the ability to unmarshal yaml
data- you can't unmarshall yaml through an opaque interface like this.

Json somehow supported it, but for yaml it triggers a panic.

Thus this: drop the KeyRequest interface and instead just rename
the concrete BasicKeyRequest struct into KeyRequest.

Many thanks to Chris Broglie for running this down including the
intended fix.