fix upgrading transitive coreos dependency breakage
The root of the problem is that the latest semver release of https://github.com/google/certificate-transparency-go is from 2018, and the etcd packages have been renamed since. For some reason the readme instructs the usage of `-u` with go get, so this upgrades that to the latest version, along with some other packages that are years out of date.

more context at: https://github.com/etcd-io/etcd/issues/11749

resolves #1142, 
resolves #1182
resolves #1183
resolves #1188