Improve IPv6 address handling in ocspserve
When setting an IPv6 address to listing on via the -address command-line
argument for both serve and ocspserve, the latter errors with "listen
tcp: address ::1:8889: too many colons in address" unless it is escaped.
However, the former uses the net library to process the address and
port, which results in the enforced escaping of IPv6 addresses
regardless of if the address is already enclosed in square brackets
(e.g. [::1]).

This changes oscpserve to use the same net library call as serve to
provide consistency between the two calls when handling IPv6 addresses.