transport/ca/localca: New(): return error instead of calling os.Exit(1)
This code was added in 56dfed7c82597110116d549dc1c990a5e24d1db7, but
now has become the only use of github.com/kisom/goutils/assert, which
previously was used in tests, and now is archived (moved to a new module).

There were a couple of issues with this code;

The `assert.NoError` appears to have a bug; it accepts optional arguments,
but those are ignored; https://github.com/kisom/goutils/blob/v1.4.3/assert/assert.go#L90-L99
In this case, it meant that the additional information to describe the
error won't be printed.

Looking at the code (https://github.com/kisom/goutils/blob/v1.4.3/assert/assert.go#L35-L45),
it defaults (`GOTRACEBACK` anything other than "crash") using `os.Exit(1)`.
While (from the description), program execution MUST be terminated, there
are some downsides to using `os.Exit` here, as it terminates execution
immediately (which is desirable), but has no way to recover. While users
should NOT use the result in this case, they still may want to catch this
error (without terminating the program as a whole, which may be problematic
if this module is used as part of a service). `os.Exit` also does not execute
pending `defer` statements, which may still be desirable to handle state cleanup.

This patch changes the function to return an error instead, allowing the
caller to handle the error.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>